fix(portal): convert dynamic routes for GitHub Pages static export
- Convert dynamic route pages to use dynamic imports with ssr:false
- Create client components for order, proposal, offering details
- Add generateStaticParams with placeholder routes
- Add custom 404 page for SPA routing fallback
- Update next.config.mjs with trailingSlash for GitHub Pages

Fixes static export build failure for routes:
- /orders/[id]
- /governance/proposals/[id]
- /provider/orders/[id]
- /marketplace/[provider]/[sequence]
