+++ portal/src/app/(customer)/marketplace/[provider]/layout.tsx
@@ -1,0 +1,8 @@
+// Layout for nested dynamic route
+export async function generateStaticParams() {
+  return [{ provider: '_' }];
+}
+
+export default function ProviderLayout({ children }: { children: React.ReactNode }) {
+  return <>{children}</>;
+}
